+++ packages/storybook-a11y-test/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+### Changed
+
+- Updated to use @axe-core/puppeteer for running tests to make it compatible with storybook 6.4.x. [[#2106](https://github.com/Shopify/quilt/pull/2106)]
 
 ## 0.3.0 - 2021-09-24